+++ noodles-sam/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 [#403]: https://github.com/zaeleus/noodles/issues/403
 
+### Fixed
+
+  * sam/io/reader/record: Only strip carriage return from CRLF newline.
+
 ## 0.86.0 - 2026-07-10
 
 ### Changed